Update org.yaml (#303)
Please consider adding @nicholas-abad to the organisation. He may not be as visibly active in the Kubeflow community as I am, but is doing plenty of heavy lifting in the backgroud. He will help me with Model Management topics, and has also (as confirmed by him in https://github.com/kubeflow/community/issues/342) volunteered to act as a calendar admin.