chore: add liavweiss to kubeflow-notebooks-security
Liav is going to be helping about with KubeFlow CNCF graduation - focusing specifically on analyzing and (where appropriate) addressing CVE and/or security issues in the `notebooks-v1` branch of `kubeflow/notebooks`.

This commit adds `liavweiss` to the `kubeflow-notebooks-security` group so he can access information within the `Security` tab on the `kubeflow/notebooks` repository.

Signed-off-by: Andy Stoneberg <astonebe@redhat.com>